Issue 1457: Move LDAPFatUtils and swap equalsDns in VMM fats
