+++ dev/com.ibm.ws.security.wim.adapter.file_fat/fat/src/com/ibm/ws/security/wim/adapter/file/fat/MultipleReposTest.java
@@ -11,15 +11,13 @@
 
 package com.ibm.ws.security.wim.adapter.file.fat;
 
+import static componenttest.topology.utils.LDAPFatUtils.assertDNsEqual;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
 import java.util.List;
-
-import javax.naming.InvalidNameException;
-import javax.naming.ldap.LdapName;
 
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
@@ -181,10 +179,7 @@ public class MultipleReposTest {
         String uniqueUserId = "uid=admin,o=defaultWIMFileBasedRealm";
         Log.info(c, "getUniqueUserId", "Checking with a valid user.");
         try {
-            equalDNs("", uniqueUserId, servlet.getUniqueUserId(user));
-        } catch (InvalidNameException e) {
-            // Do you need FFDC here? Remember FFDC instrumentation and @FFDCIgnore
-            e.printStackTrace();
+            assertDNsEqual("UniqueUserId is incorrect", uniqueUserId, servlet.getUniqueUserId(user));
         } catch (EntryNotFoundException e) {
             // Do you need FFDC here? Remember FFDC instrumentation and @FFDCIgnore
             e.printStackTrace();
@@ -278,10 +273,7 @@ public class MultipleReposTest {
         String uniqueGroupId = "cn=group1,o=defaultWIMFileBasedRealm";
         Log.info(c, "getUniqueGroupId", "Checking with a valid group.");
         try {
-            equalDNs(null, uniqueGroupId, servlet.getUniqueGroupId(group));
-        } catch (InvalidNameException e) {
-            // Do you need FFDC here? Remember FFDC instrumentation and @FFDCIgnore
-            e.printStackTrace();
+            assertDNsEqual("UniqueGroupId is incorrect", uniqueGroupId, servlet.getUniqueGroupId(group));
         } catch (EntryNotFoundException e) {
             // Do you need FFDC here? Remember FFDC instrumentation and @FFDCIgnore
             e.printStackTrace();
@@ -360,10 +352,4 @@ public class MultipleReposTest {
         assertEquals("There should only be one entry", 1, list.size());
     }
 
-    // TODO Replace
-    private void equalDNs(String msg, String dn1, String dn2) throws InvalidNameException {
-        LdapName ln1 = new LdapName(dn1);
-        LdapName ln2 = new LdapName(dn2);
-        assertEquals(msg, ln1, ln2);
-    }
 }
