+++ dev/com.ibm.ws.security.wim.adapter.ldap_fat/fat/src/com/ibm/ws/security/wim/adapter/ldap/fat/LDAPRegistryDynamicUpdateTest.java
@@ -11,12 +11,12 @@
 
 package com.ibm.ws.security.wim.adapter.ldap.fat;
 
-import static com.ibm.ws.security.wim.adapter.ldap.fat.LDAPFatUtils.assertDNsEqual;
-import static com.ibm.ws.security.wim.adapter.ldap.fat.LDAPFatUtils.createADLdapRegistry;
-import static com.ibm.ws.security.wim.adapter.ldap.fat.LDAPFatUtils.createFederatedRepository;
-import static com.ibm.ws.security.wim.adapter.ldap.fat.LDAPFatUtils.createSunLdapRegistry;
-import static com.ibm.ws.security.wim.adapter.ldap.fat.LDAPFatUtils.createTDSLdapRegistry;
-import static com.ibm.ws.security.wim.adapter.ldap.fat.LDAPFatUtils.updateConfigDynamically;
+import static componenttest.topology.utils.LDAPFatUtils.assertDNsEqual;
+import static componenttest.topology.utils.LDAPFatUtils.createADLdapRegistry;
+import static componenttest.topology.utils.LDAPFatUtils.createFederatedRepository;
+import static componenttest.topology.utils.LDAPFatUtils.createSunLdapRegistry;
+import static componenttest.topology.utils.LDAPFatUtils.createTDSLdapRegistry;
+import static componenttest.topology.utils.LDAPFatUtils.updateConfigDynamically;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
