+++ dev/com.ibm.ws.security.wim.adapter.ldap_fat/fat/src/com/ibm/ws/security/wim/adapter/ldap/fat/URAPIs_TDSLDAP_SSLTest.java
@@ -11,6 +11,7 @@
 
 package com.ibm.ws.security.wim.adapter.ldap.fat;
 
+import static componenttest.topology.utils.LDAPFatUtils.assertDNsEqual;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
@@ -19,9 +20,6 @@ import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
 import java.util.List;
-
-import javax.naming.InvalidNameException;
-import javax.naming.ldap.LdapName;
 
 import org.junit.AfterClass;
 import org.junit.Assume;
@@ -405,7 +403,7 @@ public class URAPIs_TDSLDAP_SSLTest {
         String uniqueUserId = "cn=vmmtestuser,o=ibm,c=us";
         Log.info(c, "getUniqueUserId", "Checking with a valid user.");
         //assertEquals("", uniqueUserId, servlet.getUniqueUserId(user));
-        equalDNs("Unique names should be equal ", uniqueUserId, servlet.getUniqueUserId(user));
+        assertDNsEqual("Unique names should be equal ", uniqueUserId, servlet.getUniqueUserId(user));
     }
 
     /**
@@ -675,7 +673,7 @@ public class URAPIs_TDSLDAP_SSLTest {
         String group = "vmmgrp1";
         String uniqueGroupId = "cn=vmmgrp1,o=ibm,c=us";
         Log.info(c, "getUniqueGroupId", "Checking with a valid group.");
-        equalDNs(null, uniqueGroupId, servlet.getUniqueGroupId(group));
+        assertDNsEqual("UniqueGroupId is incorrect", uniqueGroupId, servlet.getUniqueGroupId(group));
     }
 
     /**
@@ -926,10 +924,4 @@ public class URAPIs_TDSLDAP_SSLTest {
         }
     }
 
-    // TODO REPLACE
-    private void equalDNs(String msg, String dn1, String dn2) throws InvalidNameException {
-        LdapName ln1 = new LdapName(dn1);
-        LdapName ln2 = new LdapName(dn2);
-        assertEquals(msg, ln1, ln2);
-    }
 }
