+++ dev/com.ibm.ws.security.wim.registry_fat/fat/src/com/ibm/ws/security/wim/registry/fat/DefaultWIMRealmTest.java
@@ -11,15 +11,13 @@
 
 package com.ibm.ws.security.wim.registry.fat;
 
+import static componenttest.topology.utils.LDAPFatUtils.assertDNsEqual;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
 import java.util.List;
-
-import javax.naming.InvalidNameException;
-import javax.naming.ldap.LdapName;
 
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
@@ -107,8 +105,8 @@ public class DefaultWIMRealmTest {
         String user = "persona1";
         String password = "ppersona1";
         Log.info(c, "checkPassword", "Checking good credentials");
-        equalDNs("Authentication should succeed.",
-                 "uid=persona1,ou=users,dc=rtp,dc=raleigh,dc=ibm,dc=com", servlet.checkPassword(user, password));
+        assertDNsEqual("Authentication should succeed.",
+                       "uid=persona1,ou=users,dc=rtp,dc=raleigh,dc=ibm,dc=com", servlet.checkPassword(user, password));
     }
 
     /**
@@ -234,7 +232,7 @@ public class DefaultWIMRealmTest {
         String user = "persona1";
         String uniqueUserId = "uid=persona1,ou=users,dc=rtp,dc=raleigh,dc=ibm,dc=com";
         Log.info(c, "getUniqueUserId", "Checking with a valid user.");
-        equalDNs("", uniqueUserId, servlet.getUniqueUserId(user));
+        assertDNsEqual("UniqueUserId is incorrect", uniqueUserId, servlet.getUniqueUserId(user));
     }
 
     /**
@@ -441,7 +439,7 @@ public class DefaultWIMRealmTest {
         String group = "vmmgroup1";
         String uniqueGroupId = "cn=vmmgroup1,ou=users,dc=rtp,dc=raleigh,dc=ibm,dc=com";
         Log.info(c, "getUniqueGroupId", "Checking with a valid group.");
-        equalDNs(null, uniqueGroupId, servlet.getUniqueGroupId(group));
+        assertDNsEqual("UniqueGroupId is incorrect", uniqueGroupId, servlet.getUniqueGroupId(group));
     }
 
     /**
@@ -474,7 +472,7 @@ public class DefaultWIMRealmTest {
     public void getGroupSecurityName() throws Exception {
         String uniqueGroupId = "vmmgroup1";
         Log.info(c, "getGroupSecurityName", "Checking with a valid group.");
-        equalDNs("", "cn=vmmgroup1,ou=users,dc=rtp,dc=raleigh,dc=ibm,dc=com", servlet.getGroupSecurityName(uniqueGroupId));
+        assertDNsEqual("Group name is incorrect", "cn=vmmgroup1,ou=users,dc=rtp,dc=raleigh,dc=ibm,dc=com", servlet.getGroupSecurityName(uniqueGroupId));
     }
 
     /**
@@ -667,9 +665,4 @@ public class DefaultWIMRealmTest {
         assertTrue("An invalid user should cause EntryNotFoundException", true);
     }
 
-    private void equalDNs(String msg, String dn1, String dn2) throws InvalidNameException {
-        LdapName ln1 = new LdapName(dn1);
-        LdapName ln2 = new LdapName(dn2);
-        assertEquals(msg, ln1, ln2);
-    }
 }
